+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.3.0 March 15, 2024
 
+- Add `TypeCaster::toNonNegativeIntOrNull()`.
 - Raise minimal PHP version to 8.1.
 
 ## 0.2.4 June 15, 2023